change run to exec, elevate 440 to 550, add example test if everything works right
it's impossible to do `docker exec` when the passbolt container isn't
running, and the passbolt container won't run if it can't read the gpg
keys, as it keeps restarting. Thus it's better to use `run --rm`

I just had an outage because 440 permissions don't allow passbolt to
execute commands on the folder, and subsequently the files, which is why
it kepth throwing permission errors and boot looping.

the `-u www-data cat` would have been very useful to debugging this, so
I think it's helpful enough to test if user can read the file before
turning passbolt on.

I also changed the 4 direct file chown/chmods to use `-R` and apply to
the folder too, because the folder also has to be owned and accessible
to the `www-data user`.

after applying these, I could finally resume my deployment.